+++ packages/cli/test/generate-index.spec.ts
@@ -147,7 +147,7 @@ describe('build index', () => {
         expect(res.trim()).to.equal(
             [
                 '@namespace "INDEX";',
-                ':import {-st-from: "./comp-A.st.css";-st-default:CompA;-st-named: a as CompA__a, color1 as var_CompA__color1, --color2 as --CompA__color2, keyframes(X as CompA__X);}',
+                ':import {-st-from: "./comp-A.st.css";-st-default:CompA;-st-named: a as CompA__a, color1 as CompA__color1, --color2 as --CompA__color2, keyframes(X as CompA__X);}',
                 '.root CompA{}',
                 '.root .CompA__a{}',
                 ':import {-st-from: "./b/1-some-comp-B-.st.css";-st-default:SomeCompB;-st-named: b as SomeCompB__b;}',
